Removed BigDecimal due to not implementing BigDecimal Math Library

+++ app/src/main/java/nl/politie/politiebureau/domain/PolitiebureauEntity.java
@@ -4,14 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class PolitiebureauEntity {
     private String name;
     private String plaats;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
 }
